scmxpertlite_srillu_final_frm/scmxpertlite_srillu_final/backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt,argon2]
pydantic
kafka-python
python-dotenv
Jinja2
python-multipart
requests
email-validator
itsdangerous
pymongo
aiofiles